+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Development]
+### Added
+- Added `RedactClient` and the ability to interact with the Nexmo Redact API.
+
 ## [3.6.0] - 2018-07-06
 ### Added
 - Added `getSmsPrice` to `AccountClient` for getting SMS pricing for a country.