+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Added `RedactClient` and the ability to interact with the Nexmo Redact API.
 
+### Changed
+- `TalkNcco` now uses `VoiceName` internally instead of `String` for the `voiceName` property.
+
 ## [3.6.0] - 2018-07-06
 ### Added
 - Added `getSmsPrice` to `AccountClient` for getting SMS pricing for a country.
@@ -16,9 +19,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added `topUp` to `AccountClient` for topping up your account which has auto-reload enabled.
 - Added `getSms` to `SmsClient` for searching for a single message by id.
 - Added `ConversionClient` and the ability to interact with the Nexmo Conversion API.
-
-### Changed
-- `TalkNcco` now uses `VoiceName` internally instead of `String` for the `voiceName` property.
 
 ## [3.5.0] - 2018-05-29
 ### Changed